backend/requirements.txt
Flask
openai
python-dotenv
psycopg2  # If you are using PostgreSQL
Flask-SQLAlchemy